test(gotmpl4j-sprig): generic Sprig conformance suite (+numeric/defaults/list/dict)
Fold the strings-specific harness into a parameterized SprigConformanceTest driven
by category (one base64 TSV per Sprig test file), extending the grind to numeric,
defaults, list and dict — 53 no-variable runt(tpl, expect) cases from
Masterminds/sprig, rendered through gotmpl4j and compared to Sprig.

All conformant except one pinned numeric float-accumulation edge:
`{{ 1.2 | mulf "2.4" 10 "4"}}` — 2.4*10*4*1.2 is 115.19999999999999 (one ulp below
the nearest double to 115.2); gotmpl4j returns that exact IEEE value and Java's
Double.toString is correctly shortest for it, while Go's strconv prints the
adjacent 115.2. Matching it would need a Go-faithful float formatter in GoFmt
(which the 346-chart parity depends on) — pinned, not worth the risk for this edge.

Test-only. All 534 sprig tests green.
