+++ jhelm-gotemplate-sprig/src/test/resources/conformance/sprig_defaults_cases.tsv
@@ -1,0 +1,14 @@
+e3siIiB8IGRlZmF1bHQgImZvbyJ9fQ==	Zm9v
+e3tkZWZhdWx0ICJmb28iIDIzNH19	MjM0
+e3tkZWZhdWx0ICJmb28iIDIuMzR9fQ==	Mi4zNA==
+e3sgLk5vdGhpbmcgfCBkZWZhdWx0ICIxMjMiIH19	MTIz
+e3sgZGVmYXVsdCAiMTIzIiB9fQ==	MTIz
+e3tpZiBlbXB0eSAxfX0xe3tlbHNlfX0we3tlbmR9fQ==	MA==
+e3tpZiBlbXB0eSAwfX0xe3tlbHNlfX0we3tlbmR9fQ==	MQ==
+e3tpZiBlbXB0eSAiIn19MXt7ZWxzZX19MHt7ZW5kfX0=	MQ==
+e3tpZiBlbXB0eSAwLjB9fTF7e2Vsc2V9fTB7e2VuZH19	MQ==
+e3tpZiBlbXB0eSBmYWxzZX19MXt7ZWxzZX19MHt7ZW5kfX0=	MQ==
+e3t0cnVlIHwgdGVybmFyeSAiZm9vIiAiYmFyIn19	Zm9v
+e3t0ZXJuYXJ5ICJmb28iICJiYXIiIHRydWV9fQ==	Zm9v
+e3tmYWxzZSB8IHRlcm5hcnkgImZvbyIgImJhciJ9fQ==	YmFy
+e3t0ZXJuYXJ5ICJmb28iICJiYXIiIGZhbHNlfX0=	YmFy
