+++ jhelm-gotemplate-sprig/src/test/resources/conformance/sprig_dict_cases.tsv
@@ -1,0 +1,1 @@
+e3skdCA6PSBkaWN0ICJJIiAic2hvdCIgInRoZSIgImFsYmF0cm9zcyJ9fXt7JHQudGhlfX0ge3skdC5JfX0=	YWxiYXRyb3NzIHNob3Q=
