+++ jhelm-gotemplate-sprig/src/test/resources/conformance/sprig_list_cases.tsv
@@ -1,0 +1,2 @@
+e3skdCA6PSB0dXBsZSAxICJhIiAiZm9vIn19e3tpbmRleCAkdCAyfX17e2luZGV4ICR0IDAgfX17e2luZGV4ICR0IDF9fQ==	Zm9vMWE=
+e3skdCA6PSBsaXN0IDEgImEiICJmb28ifX17e2luZGV4ICR0IDJ9fXt7aW5kZXggJHQgMCB9fXt7aW5kZXggJHQgMX19	Zm9vMWE=
