+++ jhelm-gotemplate-sprig/src/test/resources/conformance/sprig_numeric_cases.tsv
@@ -1,0 +1,18 @@
+e3sgYmlnZ2VzdCAxIDIgMyAzNDUgNSA2IDd9fQ==	MzQ1
+e3sgbWF4IDM0NX19	MzQ1
+e3sgbWF4ZiAxIDIgMyAzNDUuNyA1IDYgN319	MzQ1Ljc=
+e3sgbWF4IDM0NSB9fQ==	MzQ1
+e3sgbWluIDEgMiAzIDM0NSA1IDYgN319	MQ==
+e3sgbWluIDM0NX19	MzQ1
+e3sgbWluZiAxLjQgMiAzIDM0NS42IDUgNiA3fX0=	MS40
+e3sgbWluZiAzNDUgfX0=	MzQ1
+e3sgMyB8IGFkZDEgfX0=	NA==
+e3sgMy40IHwgYWRkMWYgfX0=	NC40
+e3sgMyB8IGFkZCAxIDJ9fQ==	Ng==
+e3sgMyB8IGFkZGYgMS41IDIuMn19	Ni43
+e3sgNCB8IGRpdiA1IH19	MQ==
+e3sgMiB8IGRpdmYgNSA0IH19	MC42MjU=
+e3sgMSB8IG11bCAiMiIgMyAiNCJ9fQ==	MjQ=
+e3sgMS4yIHwgbXVsZiAiMi40IiAxMCAiNCJ9fQ==	MTE1LjI=
+e3sgMyB8IHN1YiAxNCB9fQ==	MTE=
+e3sgMyB8IHN1YmYgNC41IDEgfX0=	MC41
